Updated code
Modified section for inventory. Needed to first import into temporary
table then do a SQL SELECT into one or the other tables (Warehouse or
Non-Warehouse) with a subset of columns selected.
Passed through a caption for the dialog box for getting the filename so
that there is a guide to which file is being requested.